+++ CHANGELOG.md
@@ -18,6 +18,7 @@ See `PUBLISH.md` for instructions on how to publish a new version.
 -->
 
 - (patch) Dependency updates
+- (patch) Isolate Prism webpack logic in own file
 - (minor) Add util to restrict Prism bundle in Webpack